+++ RedResQ_API.Lib/Services/HazardService.cs
@@ -65,7 +65,7 @@ namespace RedResQ_API.Lib.Services
             throw new NotFoundException("No Hazards found!");
         }
 
-        public static async Task<bool> Add(string title, float lat, float lon, int radius, int typeId)
+        public static async Task<bool> Add(string title, double lat, double lon, int radius, int typeId)
         {
             List<SqlParameter> parameters = new List<SqlParameter>();
             string storedProcedure = "SP_Hz_AddHazard";
@@ -89,7 +89,7 @@ namespace RedResQ_API.Lib.Services
             throw new UnprocessableEntityException();
         }
 
-        public static bool Edit(long id, string? title, float? lat, float? lon, int? radius, int? typeId)
+        public static bool Edit(long id, string? title, double? lat, double? lon, int? radius, int? typeId)
         {
             List<SqlParameter> parameters = new List<SqlParameter>();
             string storedProcedure = "SP_Hz_EditHazard";
